% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_DT_cache.r
\name{clear_global_DT_cache}
\alias{clear_global_DT_cache}
\title{Clear the global DT cache}
\usage{
clear_global_DT_cache()
}
\value{
No return value, called for side effects.
}
\description{
Clear the global DT cache
}
\keyword{internal}
