% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_DT_cache.r
\name{get_cached_DT}
\alias{get_cached_DT}
\title{Get cached DT data}
\usage{
get_cached_DT(table_name, params_start, params_len)
}
\value{
list with elements: recordsTotal, recordsFiltered, cached_data
}
\description{
Return cached DT data for a table and paging range.
}
\keyword{internal}
