% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.r
\name{runExplorer}
\alias{runExplorer}
\title{Run the OMOP Explorer Shiny app}
\usage{
runExplorer(con, ...)
}
\arguments{
\item{con}{List. Database connection created by \code{duckdbCon()} or \code{mockCon()}.}

\item{...}{Additional arguments passed to \code{shiny::runApp()}.}
}
\description{
Run the OMOP Explorer Shiny app
}
\examples{
con <- mockCon()
\dontrun{
runExplorer(con)
}
DBI::dbDisconnect(con$dbcon, shutdown = TRUE)
}
